@@ -1,21 +1,22 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk.Web">
 
   <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
     <UserSecretsId>034fbc91-b434-41de-9fb1-b4d29d91a6ac</UserSecretsId>
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="5.*" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.InMemory" Version="5.*" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="5.*" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="5.*">
+    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="6.*" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.InMemory" Version="6.*" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="6.*" />
+    <PackageReference Include="Microsoft.EntityFrameworkCore.Tools" Version="6.*">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
     </PackageReference>
-    <PackageReference Include="Microsoft.Extensions.Http" Version="5.*" />
-    <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="5.*" />
-    <PackageReference Include="Microsoft.Extensions.Http.Polly" Version="5.*" />
+    <PackageReference Include="Microsoft.Extensions.Http" Version="6.*" />
+    <PackageReference Include="Microsoft.Extensions.Http.Polly" Version="6.*" />
+    <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="6.*" />
+    <PackageReference Include="Microsoft.Extensions.Logging.Debug" Version="6.*" />   
   </ItemGroup>
 
   <ItemGroup>
