@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
+    <TargetFramework>net6.0</TargetFramework>
 
     <IsPackable>false</IsPackable>
 
@@ -39,11 +39,11 @@
   </ItemGroup>
 
   <ItemGroup>
-    <PackageReference Include="Microsoft.Extensions.Configuration" Version="5.0.*" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="6.0.*" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.EnvironmentVariables" Version="5.0.*" />
-    <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="6.0.*" />
-    <PackageReference Include="Microsoft.Extensions.Configuration.UserSecrets" Version="5.0.*" />
+    <PackageReference Include="Microsoft.Extensions.Configuration" Version="6.*" />
+    <PackageReference Include="Microsoft.Extensions.Configuration.Binder" Version="6.*" />
+    <PackageReference Include="Microsoft.Extensions.Configuration.EnvironmentVariables" Version="6.*" />
+    <PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="6.*" />
+    <PackageReference Include="Microsoft.Extensions.Configuration.UserSecrets" Version="6.*" />
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="17.0.*" />
     <PackageReference Include="NSubstitute" Version="4.2.2" />
     <PackageReference Include="xunit" Version="2.4.1" />
